+++ rust-src/concordium_base/src/sigma_protocols/com_enc_eq.rs
@@ -5,6 +5,7 @@
 //! Pedersen commitment.
 
 use super::common::*;
+use crate::random_oracle::StructuredDigest;
 use crate::{
     common::*,
     curve_arithmetic::{multiexp, Curve, Field},