+++ rust-src/concordium_base/src/sigma_protocols/com_ineq.rs
@@ -14,6 +14,7 @@ use super::{
     com_mult::{ComMult, ComMultSecret, Response as ComMultResponse},
     common::{prove as sigma_prove, verify as sigma_verify, SigmaProof},
 };
+use crate::random_oracle::StructuredDigest;
 use crate::{
     common::*,
     curve_arithmetic::{Curve, Field},